remove unnecessary type conversions

+++ node/node.go
@@ -544,7 +544,7 @@ func (n *Node) initChainManager(avaxAssetID ids.ID) error {
 	n.chainManager = chains.New(&chains.ManagerConfig{
 		StakingEnabled:          n.Config.EnableStaking,
 		MaxPendingMsgs:          n.Config.MaxPendingMsgs,
-		MaxNonStakerPendingMsgs: uint32(n.Config.MaxNonStakerPendingMsgs),
+		MaxNonStakerPendingMsgs: n.Config.MaxNonStakerPendingMsgs,
 		StakerMSGPortion:        n.Config.StakerMSGPortion,
 		StakerCPUPortion:        n.Config.StakerCPUPortion,
 		Log:                     n.Log,
